+++ src/MagicGradients.Forms/Animation/ITweener.cs
