+++ src/MagicGradients.Forms/Animation/Library/DoubleAnimation.cs
@@ -1,5 +1,13 @@
 ﻿namespace MagicGradients.Forms.Animation
 {
+    public class DoubleTweener : ITweener<double>
+    {
+        public double Tween(double @from, double to, double progress)
+        {
+            return from + (to - from) * progress;
+        }
+    }
+
     public class DoubleAnimation : PropertyAnimation<double>
     {
         public override ITweener<double> Tweener { get; } = new DoubleTweener();
@@ -10,8 +18,5 @@
         public override ITweener<double> Tweener { get; } = new DoubleTweener();
     }
 
-    public class DoubleKeyFrame : KeyFrame<double>
-    {
-
-    }
+    public class DoubleKeyFrame : KeyFrame<double> { }
 }
