+++ src/MagicGradients.Forms/Animation/Library/IntegerAnimation.cs
@@ -1,5 +1,13 @@
 ﻿namespace MagicGradients.Forms.Animation
 {
+    public class IntegerTweener : ITweener<int>
+    {
+        public int Tween(int @from, int to, double progress)
+        {
+            return (int)(from + (to - from) * progress);
+        }
+    }
+
     public class IntegerAnimation : PropertyAnimation<int>
     {
         public override ITweener<int> Tweener { get; } = new IntegerTweener();
@@ -10,8 +18,5 @@
         public override ITweener<int> Tweener { get; } = new IntegerTweener();
     }
 
-    public class IntegerKeyFrame : KeyFrame<int>
-    {
-
-    }
+    public class IntegerKeyFrame : KeyFrame<int> { }
 }
