+++ src/MagicGradients.Forms/Animation/Library/PointAnimation.cs
@@ -2,6 +2,16 @@
 
 namespace MagicGradients.Forms.Animation
 {
+    public class PointTweener : ITweener<Point>
+    {
+        public Point Tween(Point @from, Point to, double progress)
+        {
+            return new Point(
+                from.X + (to.X - from.X) * progress,
+                from.Y + (to.Y - from.Y) * progress);
+        }
+    }
+
     public class PointAnimation : PropertyAnimation<Point>
     {
         public override ITweener<Point> Tweener { get; } = new PointTweener();
@@ -12,8 +22,5 @@ namespace MagicGradients.Forms.Animation
         public override ITweener<Point> Tweener { get; } = new PointTweener();
     }
 
-    public class PointKeyFrame : KeyFrame<Point>
-    {
-
-    }
+    public class PointKeyFrame : KeyFrame<Point> { }
 }
