+++ src/MagicGradients.Forms/Animation/Library/ThicknessAnimation.cs
@@ -2,6 +2,18 @@
 
 namespace MagicGradients.Forms.Animation
 {
+    public class ThicknessTweener : ITweener<Thickness>
+    {
+        public Thickness Tween(Thickness @from, Thickness to, double progress)
+        {
+            return new Thickness(
+                from.Left + (to.Left - from.Left) * progress,
+                from.Top + (to.Top - from.Top) * progress,
+                from.Right + (to.Right - from.Right) * progress,
+                from.Bottom + (to.Bottom - from.Bottom) * progress);
+        }
+    }
+
     public class ThicknessAnimation : PropertyAnimation<Thickness>
     {
         public override ITweener<Thickness> Tweener { get; } = new ThicknessTweener();
@@ -12,8 +24,5 @@ namespace MagicGradients.Forms.Animation
         public override ITweener<Thickness> Tweener { get; } = new ThicknessTweener();
     }
 
-    public class ThicknessKeyFrame : KeyFrame<Thickness>
-    {
-
-    }
+    public class ThicknessKeyFrame : KeyFrame<Thickness> { }
 }
