--- src/MagicGradients.Forms/Animation/Tween/IntegerTweener.cs
@@ -1,10 +1,0 @@
-﻿namespace MagicGradients.Forms.Animation
-{
-    public class IntegerTweener : ITweener<int>
-    {
-        public int Tween(int @from, int to, double progress)
-        {
-            return (int)(from + (to - from) * progress);
-        }
-    }
-}
